feat: Rename UserBadges to UserPrivilegeBadges; ref: Change style of UserPrivilegeBadges;
